Create README.md
o	Reminding the user important days like birthdays, anniversaries & other occasions of family & friends; with this application user can easily store their timeline details. 
o	On the occasion date, a text message will automatically send to the target user using mobile carrier. This application does not need the internet. 
o	An addition, this application contains the to-do-list with text-to-speech technology. When the given time of event approaches, the application will speak up the event. The main theme of this application is to protect the relationship between members.